validador

+++ linkCalculator.h
@@ -7,8 +7,8 @@
 
 using namespace std;
 
-double freeSpace(float Pt, float Gt, float Gr, float d, int f);
-double ricochet(float Pt, float Gt, float Gr, float Ht, float Hr, float d);
+long double freeSpace(float Pt, float Gt, float Gr, float d, int f);
+long double ricochet(float Pt, float Gt, float Gr, float Ht, float Hr, float d);
 long double toRadians(const long double degree);
 long double distance(long double lat1, long double long1, long double lat2, long double long2);
